break server and client into separate files
